Haley Holmes - Image Analysis Pipeline Edit
Edits made on 01/04/21 by Haley Holmes. Some paragraph reformatting, small grammar and spelling fixes. finFindR needs a subsection and navigational sidebar link as it is the only program excluded on this page that appears in the ID "ecosystem" diagram (the one with zebras and whales). Unfortunately I have no knowledge on this program and therefore cannot aid in this matter.